+++ frontend/client/views/events/BoardView.vue
@@ -23,45 +23,52 @@
         <div class="subtitle" v-if="boardBodies.length === 0">You are not a board member of any body.</div>
         <div class="subtitle" v-if="!selectedBody && boardBodies.length > 0">You haven't selected the antenna yet.</div>
 
-        <table class="table is-narrow is-fullwidth" v-if="selectedBody && boardBodies.length > 0">
-          <thead>
-            <tr>
-              <th>Date modified</th>
-              <th>Event</th>
-              <th>Name</th>
-              <th>Comment</th>
-              <th></th>
-            </tr>
-          </thead>
-          <tbody>
-            <tr v-for="(application, index) in applications" v-bind:key="index">
-              <td>{{ application.updated_at | datetime }}</td>
-              <td>
-                <router-link :to="{ name: 'oms.events.view', params: { id: application.event.url } }">
-                  {{ application.event.name }}
-                </router-link>
-              </td>
-              <td>
-                <router-link :to="{ name: 'oms.members.view', params: { id: application.user_id } }">
-                  {{ application.first_name }} {{ application.last_name  }}
-                </router-link>
-              </td>
-              <td class="field">
-                <div class="control">
-                  <input class="input" v-model="application.board_comment" />
-                </div>
-              </td>
-              <td>
-                <button type="button" class="button is-primary" @click="submitComment(application)">
-                  <span class="icon"><i class="fa fa-save"></i></span>
-                  <span>Save</span>
-                </button>
-              </td>
-            </tr>
-            <tr v-if="isLoading" colspan="5">Loading...</tr>
-            <tr v-if="!isLoading && applications.length === 0" colspan="5">No applications from this antenna.</tr>
-          </tbody>
-        </table>
+        <b-table
+          :data="applications"
+          :loading="isLoading"
+          v-if="selectedBody && boardBodies.length > 0">
+          <template slot-scope="props">
+            <b-table-column field="updated_at" label="Date modified" sortable>
+              {{ props.row.updated_at | datetime }}
+            </b-table-column>
+
+            <b-table-column field="event.name" label="Event" sortable>
+              <router-link :to="{ name: 'oms.events.view', params: { id: props.row.event.url } }">
+                {{ props.row.event.name }}
+              </router-link>
+            </b-table-column>
+
+            <b-table-column field="first_name" label="Name" sortable>
+              <router-link :to="{ name: 'oms.members.view', params: { id: props.row.user_id } }">
+                {{ props.row.first_name }} {{ props.row.last_name  }}
+              </router-link>
+            </b-table-column>
+
+            <b-table-column field="board_comment" label="Board comment">
+              <div class="control">
+                <input class="input" v-model="props.row.board_comment" />
+              </div>
+            </b-table-column>
+
+            <b-table-column label="Save">
+              <button type="button" class="button is-primary" @click="submitComment(props.row)">
+                <span class="icon"><i class="fa fa-save"></i></span>
+                <span>Save</span>
+              </button>
+            </b-table-column>
+          </template>
+
+          <template slot="empty">
+            <section class="section">
+              <div class="content has-text-grey has-text-centered">
+                <p>
+                  <b-icon icon="fa fa-times-circle" size="is-large"></b-icon>
+                </p>
+                <p>Nothing here.</p>
+              </div>
+            </section>
+          </template>
+        </b-table>
       </div>
     </div>
   </div>