+++ frontend/client/views/statutory/Incoming.vue
@@ -11,6 +11,8 @@
           </div>
         </div>
 
+        <div>Total participants: {{ filteredApplications.length }}</div>
+
         <b-table
           :data="filteredApplications"
           :loading="isLoading"